Delete unused stuff and cleanup
