+++ editor/src/components/canvas/ui-jsx-canvas.tsx
@@ -81,7 +81,6 @@ import {
   usePubSubAtomReadOnly,
 } from '../../core/shared/atom-with-pub-sub'
 import { omit } from '../../core/shared/object-utils'
-import { validateControlsToCheck } from './canvas-globals'
 import type { EditorDispatch } from '../editor/action-types'
 import {
   clearListOfEvaluatedFiles,
@@ -324,23 +323,6 @@ export const UiJsxCanvas = React.memo<UiJsxCanvasPropsWithErrorCallback>((props)
   resolvedFileNames.current = [uiFilePath]
   let evaluatedFileNames = React.useRef<Array<string>>([]) // evaluated (i.e. not using a cached evaluation) this render
   evaluatedFileNames.current = [uiFilePath]
-  React.useEffect(() => {
-    // FIXME: this setTimeout is a really fragile solution! `validateControlsToCheck` reads from global variables, and it is hard to guarantee that
-    // there is no race condition coming from changes from other rerenders.
-    // Now it is a little more stable because we capture the values of `resolvedFileNames` and `evaluatedFileNames`, but it would be better to make
-    // this less dependent on global variables and setTimeout.
-    const resolvedFileNamesLocal = [...resolvedFileNames.current]
-    const evaluatedFileNamesLocal = [...evaluatedFileNames.current]
-    setTimeout(() => {
-      // wrapping in a setTimeout so we don't dispatch from inside React lifecycle
-      void validateControlsToCheck(
-        dispatch,
-        propertyControlsInfo,
-        resolvedFileNamesLocal,
-        evaluatedFileNamesLocal,
-      )
-    }, 0)
-  })
 
   if (!IS_TEST_ENVIRONMENT) {
     // FIXME This is illegal! The two lines below are triggering a re-render
