+++ editor/src/components/canvas/ui-jsx.test-utils.tsx
@@ -122,7 +122,6 @@ import { createTestProjectWithCode } from '../../sample-projects/sample-project-
 import { DummyPersistenceMachine } from '../editor/persistence/persistence.test-utils'
 import type { BuiltInDependencies } from '../../core/es-modules/package-manager/built-in-dependencies-list'
 import { createBuiltInDependenciesList } from '../../core/es-modules/package-manager/built-in-dependencies-list'
-import { clearAllRegisteredControls } from './canvas-globals'
 import { createEmptyStrategyState } from './canvas-strategies/interaction-state'
 import type { DomWalkerMutableStateData } from './dom-walker'
 import {
@@ -334,7 +333,6 @@ export async function renderTestEditorWithModel(
 
   // Reset canvas globals
   resetDispatchGlobals()
-  clearAllRegisteredControls()
   clearListOfEvaluatedFiles()
 
   const asyncTestDispatch = async (
